refactored classes up to handling exceptions

+++ src/main/java/com/abm/moviesapi/controller/DirectorRestController.java
@@ -1,17 +1,8 @@
 package com.abm.moviesapi.controller;
 
-import com.abm.moviesapi.entity.Director;
-import com.abm.moviesapi.service.DirectorService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 import com.abm.moviesapi.dto.DirectorDTO;
-import com.abm.moviesapi.entity.Casting;
 import com.abm.moviesapi.entity.Director;
-import com.abm.moviesapi.entity.Genre;
+import com.abm.moviesapi.exceptions.CustomDirectorException.DirectorNotFoundException;
 import com.abm.moviesapi.service.DirectorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -37,8 +28,10 @@ public class DirectorRestController {
 
     //GET (all movies for director id)
     @GetMapping("/directors/{directorId}")
-    public DirectorDTO directorDTO(@PathVariable int directorId) throws Exception {
-
+    public DirectorDTO directorDTO(@PathVariable int directorId){
+        if (directorId <= 0 || directorId > directorService.findAll().size()){
+            throw new DirectorNotFoundException();
+        }
         Director director = directorService.findById(directorId);
         DirectorDTO directorDTO= new DirectorDTO(director.getId(), director.getDirectorName(), directorService.getMoviesByDirector(directorId));
         return  directorDTO;
@@ -46,7 +39,7 @@ public class DirectorRestController {
 
     //POST (add director)
     @PostMapping(value = "/directors", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public Director addDirector(@RequestBody Director director) throws Exception {
+    public Director addDirector(@RequestBody Director director){
         director.setId(0);
         directorService.save(director);
         return director;
@@ -54,7 +47,10 @@ public class DirectorRestController {
 
     //PUT (update director)
     @PutMapping("/directors")
-    public Director updateDirector(@RequestBody Director director) throws Exception {
+    public Director updateDirector(@RequestBody Director director){
+        if (director.getId() <= 0 || director.getId() > directorService.findAll().size()){
+            throw new DirectorNotFoundException();
+        }
         directorService.save(director);
         return director;
         //TODO
@@ -64,6 +60,9 @@ public class DirectorRestController {
     @DeleteMapping("/directors/{directorId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public HttpStatus deleteDirector(@PathVariable int directorId) throws Exception {
+        if (directorId <= 0 || directorId > directorService.findAll().size()){
+            throw new DirectorNotFoundException();
+        }
         directorService.deleteById(directorId);
         return HttpStatus.NO_CONTENT;
     }
